Include instructions for initial recommender setup
The instructions to setup the recommendation engine were created in markdown
mostly because I have really enjoyed learning to use markdown. The instruction
s are still incomplete but at their completion are supposed to be able to
guide a noob to be able to use this recommender for their own data.
The setup instructions also have embedded in them a tutorial which is
basically the setup of hiije for a supermarket application using data which
is publicly available on the interweb (Stedy's Github repository link in the
hiije readme).