Moved the entire repository from Windows to Linux and also made it a package
Moving the entire repository to Linux involved removing Windows specific files
and modifying the virtual environment. This may or may not affect the files
on the repository (it definitely affects several files on my local machine).

Making it a package involved reordering the files so that the structure of the
project becomes:

hiije/
	hiije/
		|___Core/
		|___templates/
		|___tests/
		|_____init__.py
	setup.py
	README.md
	hiije_architecture.md

After doing this it was necessary to run pip install .e . to install the
 project as a package. I write this much obvious detail because this is
something I really struggled to be able to do.

Running the unit tests at this point works. However, note that the test suite
at this point only contains a test to show that sqlalchemy is able to access
the project DB(which may even be empty!). The next few commits will be
dedicated to having functionality to populate the relevant DB's and to
(finally) start getting recommendations on Linux. Can already smell...
 Hiije_Linux